third_party/WebKit/LayoutTests/fast/dom/shadow/pseudo-not.html

<!DOCTYPE html>

<script src="../../../resources/js-test.js"></script>

<div id="host">
    <span class="first"></span>
    <span class="second"></span>
</div>

<script>
description("Test the :not pseudo selector in <content select>");

var host = document.getElementById("host");
var root = host.createShadowRoot();
var content = document.createElement("content");
root.appendChild(content);

content.select = ":not(.first)";
shouldBe("content.getDistributedNodes().length", "1");
shouldBeEqualToString("content.getDistributedNodes()[0].className", "second");

content.select = ":not(.second)";
shouldBe("content.getDistributedNodes().length", "1");
shouldBeEqualToString("content.getDistributedNodes()[0].className", "first");
</script>
