third_party/WebKit/LayoutTests/fast/dom/shadow/pseudo-not-expected.txt

CONSOLE WARNING: line 14: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Test the :not pseudo selector in

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS content.getDistributedNodes().length is 1
PASS content.getDistributedNodes()[0].className is "second"
PASS content.getDistributedNodes().length is 1
PASS content.getDistributedNodes()[0].className is "first"
PASS successfullyParsed is true

TEST COMPLETE
